requirements.txt
Flask
tensorflow
Pillow
requests
numpy
setuptools
wheel
gunicorn
google-cloud-storage==2.1.0